Retrieve Supported Metabolite Identifier Conversion Systems

This function returns the available metabolite identifier conversion systems supported by different online services, including **CTS FiehnLab**, **ChemSpider**, and **OpenAI**. It provides a list of identifier formats that can be used for conversion.

request_metabolite_id_systems(
  server = c("cts.fiehnlab", "chemspider", "openai"),
  source_format = c("local", "online")
)

Arguments

server

Character. The server to retrieve supported identifier formats from. Available options: - `"cts.fiehnlab"`: The **Chemical Translation Service (CTS) by FiehnLab** at UC Davis. - `"chemspider"`: The **ChemSpider API** provided by the Royal Society of Chemistry. - `"openai"`: Uses OpenAI's language models for metabolite ID conversions.

source_format

Character. Specifies whether to use a **local** or **online** source for fetching the available databases. Options: - `"local"`: Uses locally stored information about supported databases. - `"online"`: Fetches real-time data from the respective API endpoints.

Value

A data frame or vector containing the supported **From-To** identifier conversions for the selected `server`. - If `server = "cts.fiehnlab"`, returns a vector of available **from** and **to** databases. - If `server = "chemspider"`, returns a data frame listing valid identifier conversions. - If `server = "openai"`, returns a predefined set of supported sources.

Details

- **CTS FiehnLab**: The function queries `http://cts.fiehnlab.ucdavis.edu/rest/fromValues` to retrieve the list of supported database formats. - **ChemSpider**: Retrieves a predefined set of identifier conversions. - **OpenAI**: Returns a list of supported metabolite ID formats available through OpenAI models.

**ChemSpider API Key Requirement**: If using the `"chemspider"` server, an API key is required for querying ChemSpider’s services.

**How to Obtain a ChemSpider API Key:** 1. Visit the **Royal Society of Chemistry (RSC) Developer Portal**: [https://developer.rsc.org/](https://developer.rsc.org/) 2. Sign in or create an account. 3. Navigate to **My Account** > **API Keys**. 4. Apply for access to the **ChemSpider API**. 5. Once approved, generate your API key and copy it for use.

Examples

# Get supported conversion systems for CTS FiehnLab (local)
request_metabolite_id_systems(server = "cts.fiehnlab", source_format = "local")
#>   [1] "AAA Chemistry"                                                
#>   [2] "ABBLIS Chemicals"                                             
#>   [3] "Abbott Labs"                                                  
#>   [4] "ABI Chem"                                                     
#>   [5] "AbMole Bioscience"                                            
#>   [6] "Acesobio"                                                     
#>   [7] "Achemica"                                                     
#>   [8] "Acorn PharmaTech"                                             
#>   [9] "Active Biopharma"                                             
#>  [10] "Adooq BioScience"                                             
#>  [11] "AK Scientific"                                                
#>  [12] "AKos Consulting & Solutions"                                  
#>  [13] "Alagar Yadav"                                                 
#>  [14] "ALDRICH"                                                      
#>  [15] "Alinda Chemical"                                              
#>  [16] "Alsachim"                                                     
#>  [17] "Amadis Chemical"                                              
#>  [18] "Amatye"                                                       
#>  [19] "Ambinter"                                                     
#>  [20] "Ambit Biosciences"                                            
#>  [21] "AmicBase - Antimicrobial Activities"                          
#>  [22] "Angene Chemical"                                              
#>  [23] "Angene International"                                         
#>  [24] "Anitha"                                                       
#>  [25] "Annamalai Universtiy"                                         
#>  [26] "Annker Organics"                                              
#>  [27] "Anward"                                                       
#>  [28] "Apeiron Synthesis"                                            
#>  [29] "ApexBio Technology"                                           
#>  [30] "Apexmol"                                                      
#>  [31] "Ark Pharm"                                                    
#>  [32] "Aromsyn catalogue"                                            
#>  [33] "Aronis"                                                       
#>  [34] "ASINEX"                                                       
#>  [35] "Aurora Fine Chemicals LLC"                                    
#>  [36] "Aurum Pharmatech LLC"                                         
#>  [37] "Avanti Polar Lipids"                                          
#>  [38] "Beijing Advanced Technology Co"                               
#>  [39] "Bertin Pharma"                                                
#>  [40] "Bharathiar University"                                        
#>  [41] "Bharathidasan University"                                     
#>  [42] "Bhaskar Lab"                                                  
#>  [43] "BIDD"                                                         
#>  [44] "BIND"                                                         
#>  [45] "BindingDB"                                                    
#>  [46] "BioChemPartner"                                               
#>  [47] "BioCyc"                                                       
#>  [48] "Biological Magnetic Resonance Data Bank (BMRB)"               
#>  [49] "Bioprocess Technology Lab"                                    
#>  [50] "Biosynth"                                                     
#>  [51] "Broad Institute"                                              
#>  [52] "BroadPharm"                                                   
#>  [53] "Burnham Center for Chemical Genomics"                         
#>  [54] "Calbiochem"                                                   
#>  [55] "Calicut University"                                           
#>  [56] "Cambridge Crystallographic Data Centre"                       
#>  [57] "CAPOT"                                                        
#>  [58] "CAS"                                                          
#>  [59] "Cayman Chemical"                                              
#>  [60] "CC_PMLSC"                                                     
#>  [61] "Center for Chemical Genomics"                                 
#>  [62] "ChEBI"                                                        
#>  [63] "ChemBank"                                                     
#>  [64] "Chembase.cn"                                                  
#>  [65] "ChEMBL"                                                       
#>  [66] "ChemBlock"                                                    
#>  [67] "Chembo"                                                       
#>  [68] "ChemBridge"                                                   
#>  [69] "ChemDB"                                                       
#>  [70] "ChemExper Chemical Directory"                                 
#>  [71] "ChemFrog"                                                     
#>  [72] "Chemical Biology Department"                                  
#>  [73] "Chemical Name"                                                
#>  [74] "chemicalize.org by ChemAxon"                                  
#>  [75] "ChemIDplus"                                                   
#>  [76] "ChemMol"                                                      
#>  [77] "ChemScene"                                                    
#>  [78] "ChemSpider"                                                   
#>  [79] "ChemSynthesis"                                                
#>  [80] "ChemTik"                                                      
#>  [81] "Chiralblock Biosciences"                                      
#>  [82] "Circadian Research"                                           
#>  [83] "CLRI (CSIR)"                                                  
#>  [84] "CMLD-BU"                                                      
#>  [85] "Columbia University Molecular Screening Center"               
#>  [86] "Comparative Toxicogenomics Database"                          
#>  [87] "Creasyn Finechem"                                             
#>  [88] "Department of Bioinformatics"                                 
#>  [89] "Department of Biotechnology"                                  
#>  [90] "Department of Environmental Biotechnology"                    
#>  [91] "Department of Microbiology"                                   
#>  [92] "Department of Pharmacy"                                       
#>  [93] "Department of Zoology"                                        
#>  [94] "DiscoveryGate"                                                
#>  [95] "DrugBank"                                                     
#>  [96] "DTP/NCI"                                                      
#>  [97] "EDASA Scientific Compounds June 2013"                         
#>  [98] "EMD Biosciences"                                              
#>  [99] "Emory University Molecular Libraries Screening Center"        
#> [100] "Enamine"                                                      
#> [101] "Ennopharm"                                                    
#> [102] "EPA DSSTox"                                                   
#> [103] "Excenen Pharmatech"                                           
#> [104] "Exchemistry"                                                  
#> [105] "Faculty of Marine Sciences"                                   
#> [106] "FINETECH"                                                     
#> [107] "Finley and King Labs"                                         
#> [108] "FLUKA"                                                        
#> [109] "ForeChem"                                                     
#> [110] "Fragmenta"                                                    
#> [111] "Georganics"                                                   
#> [112] "GlaxoSmithKline (GSK)"                                        
#> [113] "GLIDA"                                                        
#> [114] "GNF / Scripps Winzeler lab"                                   
#> [115] "Golm Metabolome Database (GMD)"                               
#> [116] "GPCR-Ligand Database"                                         
#> [117] "Hangzhou APIChem Technology"                                  
#> [118] "Hangzhou Trylead Chemical Technology"                         
#> [119] "Harvard Medical School"                                       
#> [120] "HDH Pharma"                                                   
#> [121] "Human Metabolome Database"                                    
#> [122] "HUMGENEX"                                                     
#> [123] "IBCH RAS"                                                     
#> [124] "IBM"                                                          
#> [125] "ICCB-Longwood/NSRB Screening Facility"                        
#> [126] "IIT Guwahati"                                                 
#> [127] "Immunology Lab"                                               
#> [128] "Inc."                                                         
#> [129] "Inc. (AKSCI)"                                                 
#> [130] "InChIKey"                                                     
#> [131] "InFarmatik"                                                   
#> [132] "Inhibitor 2"                                                  
#> [133] "Insect Molecular Biology Lab"                                 
#> [134] "Ion Channels and Transporters"                                
#> [135] "IS Chemical Technology"                                       
#> [136] "Isoprenoids"                                                  
#> [137] "iThemba Pharmaceuticals"                                      
#> [138] "IUPHAR-DB"                                                    
#> [139] "Jamson Pharmachem Technology"                                 
#> [140] "Japan Chemical Substance Dictionary (Nikkaji)"                
#> [141] "Johns Hopkins Ion Channel Center"                             
#> [142] "Karpagam University"                                          
#> [143] "Kay Laboratory"                                               
#> [144] "KEGG"                                                         
#> [145] "Kingston Chemistry"                                           
#> [146] "KUMGM"                                                        
#> [147] "LeadScope"                                                    
#> [148] "Leiden University Medical Center"                             
#> [149] "LipidMAPS"                                                    
#> [150] "LLC"                                                          
#> [151] "LMSD"                                                         
#> [152] "LMU"                                                          
#> [153] "Ltd"                                                          
#> [154] "M.Jeyam and G.Shalini. Biochematics Division"                 
#> [155] "Marine Medicinal Plant Biotechnology Laboratory"              
#> [156] "Max Planck Institute of Molecular Physiology"                 
#> [157] "Max Planck Institute of Molecular Plant Physiology"           
#> [158] "MedChemexpress MCE"                                           
#> [159] "MIC Scientific"                                               
#> [160] "MICAD"                                                        
#> [161] "Milwaukee Institute for Drug Discovery"                       
#> [162] "MLSMR"                                                        
#> [163] "MMDB"                                                         
#> [164] "Molecular Libraries Program"                                  
#> [165] "MOLI"                                                         
#> [166] "MolPort"                                                      
#> [167] "MP Biomedicals"                                               
#> [168] "MTDP"                                                         
#> [169] "Nanjing Pharmaceutical Factory"                               
#> [170] "Nantong Baihua Bio-Pharmaceutical Co."                        
#> [171] "National Cancer Institute (NCI)"                              
#> [172] "Nature Chemical Biology"                                      
#> [173] "Nature Chemistry"                                             
#> [174] "Nature Communications"                                        
#> [175] "NCGC"                                                         
#> [176] "NIAID"                                                        
#> [177] "NIH Clinical Collection"                                      
#> [178] "NINDS Approved Drug Screening Program"                        
#> [179] "NIST"                                                         
#> [180] "NIST Chemistry WebBook"                                       
#> [181] "Nitric Oxide Research"                                        
#> [182] "NMMLSC"                                                       
#> [183] "NMRShiftDB"                                                   
#> [184] "NovoSeek"                                                     
#> [185] "Oakwood Products"                                             
#> [186] "ORST SMALL MOLECULE SCREENING CENTER"                         
#> [187] "P.Ravikumar"                                                  
#> [188] "P3 BioSystems"                                                
#> [189] "PANACHE"                                                      
#> [190] "Paul Baures"                                                  
#> [191] "PCMD"                                                         
#> [192] "PDSP"                                                         
#> [193] "PENN-ABS"                                                     
#> [194] "PennChem-GAM"                                                 
#> [195] "PFC"                                                          
#> [196] "Phytomatics Laboratory"                                       
#> [197] "priyadharshini sabarathinam angayarkanni murugesh palaniswamy"
#> [198] "Prous Science Drugs of the Future"                            
#> [199] "PubChem CID"                                                  
#> [200] "Pubchem SID"                                                  
#> [201] "Quorum sensing and Peptidomimetics Laboratory"                
#> [202] "R.Sathishkumar"                                               
#> [203] "R&D Chemicals"                                                
#> [204] "Rangan Lab"                                                   
#> [205] "RSChem"                                                       
#> [206] "S.GURUDEEBAN"                                                 
#> [207] "SASTRA University"                                            
#> [208] "SCRIPDB"                                                      
#> [209] "Selleck Chemicals"                                            
#> [210] "Selleckbio"                                                   
#> [211] "SGCOxCompounds"                                               
#> [212] "SGCStoCompounds"                                              
#> [213] "Shanghai Institute of Organic Chemistry"                      
#> [214] "Shanghai Sinofluoro Scientific Company"                       
#> [215] "SIGMA"                                                        
#> [216] "Sigma-Aldrich"                                                
#> [217] "SLING Consortium"                                             
#> [218] "SMID"                                                         
#> [219] "SMILES"                                                       
#> [220] "Southern Research Institute"                                  
#> [221] "Southern Research Specialized Biocontainment Screening Center"
#> [222] "Specialized Chemistry Center"                                 
#> [223] "Specs"                                                        
#> [224] "Sri Venkateswara University"                                  
#> [225] "SRMLSC"                                                       
#> [226] "Structural Genomics Consortium"                               
#> [227] "SureChem"                                                     
#> [228] "SYNCHEM OHG"                                                  
#> [229] "Syntechem"                                                    
#> [230] "T.RAMANATHAN & K.SATYAVANI"                                   
#> [231] "TCI (Tokyo Chemical Industry)"                                
#> [232] "ten Dijke Lab"                                                
#> [233] "Tetrahedron Scientific Inc"                                   
#> [234] "The Scripps Research Institute Molecular Screening Center"    
#> [235] "Therapeutic Targets Database"                                 
#> [236] "Thomson Pharma"                                               
#> [237] "TimTec"                                                       
#> [238] "Total TOSLab Building-Blocks"                                 
#> [239] "Tox21"                                                        
#> [240] "True PharmaChem"                                              
#> [241] "Tyger Scientific"                                             
#> [242] "UCLA Molecular Screening Shared Resource"                     
#> [243] "UM-BBD"                                                       
#> [244] "UniCarbKB"                                                    
#> [245] "University of California at San Diego (UCSD)"                 
#> [246] "University of Kansas"                                         
#> [247] "University of Michigan"                                       
#> [248] "University of Pittsburgh Molecular Library Screening Center"  
#> [249] "UPCMLD"                                                       
#> [250] "Vanderbilt Screening Center for GPCRs"                        
#> [251] "Vanderbilt Specialized Chemistry Center"                      
#> [252] "Vanderbilt University Medical Center"                         
#> [253] "VIT University"                                               
#> [254] "Vitas-M Laboratory"                                           
#> [255] "Watec Laboratories"                                           
#> [256] "Watson International Ltd"                                     
#> [257] "Web of Science"                                               
#> [258] "xPharm"                                                       
#> [259] "Zancheng Functional Chemicals"                                
#> [260] "zealing chemical"                                             
#> [261] "ZINC"